requirements.txt
aiofiles==22.1.0
anyio==3.6.2
asgi-csrf==0.9
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-default-group-wheel==1.2.2
datasette==0.63.1
datasette-saved-queries==0.2.1
datasette-vega==0.6.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
hupper==1.10.3
idna==3.4
itsdangerous==2.1.2
janus==1.0.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mergedeep==1.3.4
mypy-extensions==0.4.3
pathspec==0.10.2
Pint==0.20.1
platformdirs==2.5.4
pluggy==1.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlite-fts4==1.0.3
sqlite-utils==3.30
tabulate==0.9.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
